requirements.txt
# ── 필수 경량 (부팅 즉시 필요) ─────────────────────────────────────────
streamlit>=1.35.0,<2.0.0
streamlit-js-eval>=0.1.7,<1.0.0
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.32.0,<0.35.0
cryptography>=42.0.0,<45.0.0
pyotp>=2.9.0,<3.0.0
rapidfuzz>=3.0.0,<4.0.0
ftfy>=6.1.1,<7.0.0
# ── AI / DB (Lazy Load — 첫 사용 시 로드) ──────────────────────────────
google-genai>=0.3.0,<2.0.0
google-generativeai>=0.3.0,<1.0.0
supabase>=2.3.0,<3.0.0
google-cloud-documentai>=2.20.0,<3.0.0
google-cloud-storage>=2.10.0,<3.0.0
google-auth>=2.20.0,<3.0.0
# ── 문서 처리 (Lazy Load) ───────────────────────────────────────────────
Pillow>=10.0.0,<12.0.0
pdfplumber>=0.9.0,<1.0.0
pymupdf>=1.24.0,<2.0.0
python-docx>=0.8.11,<2.0.0
pypdf>=4.0.0,<6.0.0
PyPDF2>=3.0.0,<4.0.0
reportlab>=4.0.0,<5.0.0
openpyxl>=3.1.0,<4.0.0
openai>=1.0.0,<2.0.0
# ── 데이터 분석 (Lazy Load) ─────────────────────────────────────────────
pandas>=1.3.0,<3.0.0
numpy>=1.24.0,<3.0.0
opencv-python-headless>=4.8.0,<5.0.0
# ── 임베딩 (Lazy Load — requirements 제외, 사용 시점에 자동 설치) ────────
# sentence-transformers, faiss-cpu: torch 의존성으로 HF Space OOM 유발
# → modules/smart_scanner.py 내 lazy import로만 사용
